date: update index for `NULL` with `1970-01-01` to support `index_column_diff`
GitHub: GH-789

The current implementation ignores `NULL`. It means that we don't update
index for `NULL`. But it causes `index_column_diff` false positive.
`DATE` with `NULL` is processed as `1970-01-01` in Groonga.
Because Groonga uses the default value for
`NULL`. `index_column_diff` uses `1970-01-01` for `NULL` in
the expected posting lists. It causes false positive.

If we use `1970-01-01` instead of ignoring for `NULL`, we can
avoid the false positive. It also enables that we can search `NULL`
column value record by `1970-01-01`.

In general, it'll be useful but this is an incompatible change.
But it'll be acceptable because `NULL` behavior is undefined by definition.